+++ include/Ark/VM/VM.hpp
@@ -22,6 +22,8 @@ namespace Ark
         public:
             VM(bool debug=false);
 
+            void setLoadPath(const std::string& load_path);
+
             void feed(const std::string& filename);
             void feed(const bytecode_t& bytecode);
             void run();
@@ -41,10 +43,13 @@ namespace Ark
             std::vector<std::string> m_symbols;
             uint16_t m_dotc_idx;
             std::vector<Value> m_constants;
+            std::vector<std::string> m_plugins;
             std::vector<bytecode_t> m_pages;
 
             std::vector<Frame> m_frames;
             std::optional<Frame> m_saved_frame;
+
+            std::string m_load_path;
 
             void configure();
             void initFFI();