+++ src/Compiler/Instruction.cpp
@@ -5,7 +5,7 @@ namespace Ark
     namespace Compiler
     {
         Inst::Inst(Instruction inst) :
-            inst((uint8_t) inst)
+            inst(static_cast<uint8_t>(inst))
         {}
 
         Inst::Inst(uint8_t inst) :